Move CommandsTests to its test target
